pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "openapi-camel-caser"
version = "0.1.0"
description = "Transform openapi.json into openapi.yaml with camelCased operationId"
authors = ["Albert Shaidullin <3litac@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"
pre-commit = "^3.5.0"
pytest = "^7.4.2"
types-pyyaml = "^6.0.12.12"

[tool.black]
skip-string-normalization = true

[tool.isort]
profile = "black"
lines_after_imports = 2

[tool.mypy]
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
ignore_missing_imports = true
warn_unused_ignores = true
